Backup data files from lidar to USB storage
Make sure to format USB device as ext4 and put a filesystem on it. The
code expects the USB to be `bikedata` and mounts to
`/media/vti/bikedata` on the RPi.

I used Disk Utility on Fedora and once I formatted the USB disk I
created a partition called `bikedata` on it (ext4 type). That's it.