Add drop-in substitute for rpi-gpio
Needed to install lgpio==0.2.2.0 manually because there's an issue doing
it with `uv` or `pip3`. It is a dependency of `rpi-lgpio` which is the
drop-in substitute for the imported `rpi-gpio`.

Current set-up is that `lgpio v0.2.2.0` is installed manually `make` &&
`make install` on the system. Then `rpi-gpio` is added on top of it as
usual.